refactor: BlackJackResult 별도 파일로 분리
