+++ src/main/kotlin/blackjack/domain/result/BlackJackManager.kt
@@ -28,19 +28,3 @@ class BlackJackManager(
         }
     }
 }
-
-data class BlackJackResult(
-    val dealerResult: DealerResult,
-    val playerResults: List<PlayerResult>
-)
-
-data class DealerResult(
-    val win: Int,
-    val draw: Int,
-    val lose: Int
-)
-
-data class PlayerResult(
-    val playerName: String,
-    val match: Match
-)
